requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
google-auth==2.38.0
google-genai==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pypdfium2==4.30.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
